Backend API/requirements.txt
uvicorn==0.22.0
fastapi==0.96.0
pymongo==4.3.3
python-dotenv==0.21.0